remove includeMdxInWorkbench and improve flickering tools when switching stories
