+++ examples/docs/src/Configuration.stories.mdx
@@ -185,14 +185,3 @@ export const config: FwooshConfig = {
   basePath: "/docs",
 };
 ```
-
-### `includeMdxInWorkbench`
-
-By default we don't show the docs in workbench mode.
-If you want to see the mdx content in workbench mode set this to true
-
-```ts fwoosh.config.ts
-export const config: FwooshConfig = {
-  includeMdxInWorkbench: true,
-};
-```
