+++ packages/app/src/components/Workbench.tsx
@@ -7,18 +7,22 @@ import {
   Spinner,
   HeaderBar,
   HeaderTitle,
+  ErrorBoundary,
 } from "@fwoosh/components";
 import { styled } from "@fwoosh/styling";
 import { Outlet } from "react-router-dom";
 import { useId } from "@radix-ui/react-id";
-import { ParameterContext } from "@fwoosh/hooks";
+import { ParameterContext, useStoryId } from "@fwoosh/hooks";
 import { config } from "@fwoosh/app/config";
 import { CONTENT_ID } from "@fwoosh/utils";
+import { Panel, PanelGroup, PanelResizeHandle } from "react-resizable-panels";
+import { panels } from "@fwoosh/app/ui";
 
 import { StoryIdContext } from "./Story";
 import { WorkbenchSidebarTree } from "./sidebar/WorkbenchSidebarTree";
 import { useParameters } from "../hooks/useParameters";
 import { WorkbenchToolbar } from "./WorkbenchToolbar";
+import { ToolPanels } from "./ToolPanels";
 
 const StoryWrapper = styled("div", {
   position: "relative",
@@ -27,15 +31,73 @@ const StoryWrapper = styled("div", {
   flexDirection: "column",
 });
 
+const Scrollable = styled("div", {
+  overflow: "auto",
+});
+
+const PanelResizer = styled("div", {
+  width: "100%",
+  borderTopWidth: "$sm",
+  borderTopStyle: "$solid",
+  borderTopColor: "transparent",
+  zIndex: 100,
+  position: "relative",
+
+  "&:after": {
+    height: 12,
+    position: "absolute",
+    transform: "translateY(-50%)",
+    left: 0,
+    right: 0,
+    top: "50%",
+    content: "''",
+  },
+
+  "&:hover": {
+    borderColor: "$gray10",
+  },
+});
+
+const PanelContainer = styled("div", {
+  height: "100%",
+  borderTopWidth: "$sm",
+  borderTopStyle: "$solid",
+  borderTopColor: "$gray4",
+  backgroundColor: "$gray0",
+});
+
 const StoryWrapperWithParams = () => {
   const parameters = useParameters();
+  const storyId = useStoryId();
+
+  let content = (
+    <ErrorBoundary key={storyId}>
+      <Outlet />
+    </ErrorBoundary>
+  );
+
+  if (panels.length > 0 && storyId) {
+    content = (
+      <PanelGroup autoSaveId="fwoosh-tools" direction="vertical">
+        <Panel>
+          <Scrollable>{content}</Scrollable>
+        </Panel>
+        <PanelResizeHandle>
+          <PanelResizer />
+        </PanelResizeHandle>
+        <Panel collapsible maxSize={75}>
+          <PanelContainer>
+            <ToolPanels storySlug={storyId} />
+          </PanelContainer>
+        </Panel>
+      </PanelGroup>
+    );
+  }
 
   return (
     <ParameterContext.Provider value={parameters}>
       <Content id={CONTENT_ID}>
-        <StoryWrapper>
-          <Outlet />
-        </StoryWrapper>
+        <StoryWrapper>{content}</StoryWrapper>
       </Content>
     </ParameterContext.Provider>
   );
