+++ packages/hooks/src/useParams.ts
@@ -14,8 +14,8 @@ export const ParameterContext = React.createContext<
   StoryParameters | undefined
 >(undefined);
 
-export function useParameters<T = StoryParameters>() {
-  return React.useContext(ParameterContext) as T;
+export function useParameters<T extends StoryParameters>() {
+  return React.useContext(ParameterContext) as T | undefined;
 }
 
 export function useIsWorkbench() {
