+++ packages/types/src/index.ts
@@ -222,11 +222,6 @@ export interface FwooshOptions {
    * // You'll need to set this to "/docs"
    */
   basename?: string;
-  /**
-   * By default we don't show the docs in workbench mode.
-   * If you want to see the mdx content in workbench mode set this to true
-   */
-  includeMdxInWorkbench?: boolean;
   /** Theme to use for highlighting code */
   syntaxTheme?: SyntaxTheme;
 }
