+++ typings/fwoosh.d.ts
@@ -10,10 +10,7 @@ declare module "@fwoosh/app/overrides";
 
 declare module "@fwoosh/app/config" {
   import { FwooshOptions, ThemeObject } from "@fwoosh/types";
-  export const config: Pick<
-    FwooshOptions,
-    "title" | "includeMdxInWorkbench"
-  > & {
+  export const config: Pick<FwooshOptions, "title"> & {
     themes: ThemeObject[];
   };
 }
